More CPU instruction handler fixes + fix stupid incorrect data type usage when reading from banked ROM which was causing an overflow error

+++ src/Cartridge.cpp
@@ -182,6 +182,6 @@ bool Cartridge::isMegCartridge() {
     return megCartridge;
 }
 
-unsigned char Cartridge::read(unsigned short location) {
+unsigned char Cartridge::read(unsigned long location) {
     return cartridgeData[location];
 }
